+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+### Added
+
+- Create a Banner react component
+
 ### Changed
 
 - Improve semantic of glimpses
@@ -17,7 +21,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Fixed a layout issue on course glimpse
   when organization title and course title are too long
 - Add missing language parameter in LTI requests issued by LTI consumer plugin
-
 
 ## [2.11.0] - 2022-01-04